🍀 Mission 2 : gérer les commandes 📜 Commandes de livres, fonctionne mais a debug, commenter et refactoriser

+++ MediaTekDocuments/model/Commande.cs
@@ -7,9 +7,9 @@ namespace MediaTekDocuments.model
     {
         public string Id { get; }
         public DateTime DateCommande { get; }
-        public float Montant { get; }
+        public double Montant { get; }
 
-        public Commande(string id, DateTime dateCommande, float montant)
+        public Commande(string id, DateTime dateCommande, double montant)
         {
             Id = id;
             DateCommande = dateCommande;
